+++ src/common/flags.ts
@@ -32,6 +32,7 @@ export const T_MAPPED         = 'm';
 export const T_TRUE           = '1';
 export const T_FALSE          = '0';
 export const T_CALLSITE       = 'c';
+export const T_STAND_IN        = '5';
 
 export type RtSimpleType<T>   = { TΦ: T };
 export type RtVoidType        = RtSimpleType<typeof T_VOID>;