+++ src/transformer/metadata-encoder.ts
@@ -41,7 +41,30 @@ export class MetadataEncoder {
     }
 
     class(klass : ts.ClassDeclaration | ts.InterfaceDeclaration, details : ClassDetails) {
+        let type = this.checker.getTypeAtLocation(klass);
+
         let decs : ts.Decorator[] = [
+            ts.factory.createDecorator(
+                ts.factory.createArrowFunction(
+                    [], [], 
+                    [
+                        ts.factory.createParameterDeclaration([], [], undefined, 't')
+                    ], 
+                    undefined, 
+                    ts.factory.createToken(ts.SyntaxKind.EqualsGreaterThanToken),
+                    ts.factory.createBinaryExpression(
+                        ts.factory.createElementAccessExpression(
+                            ts.factory.createPropertyAccessExpression(
+                                ts.factory.createIdentifier('__RΦ'),
+                                't'
+                            ),
+                            type['id']
+                        ),
+                        ts.factory.createToken(ts.SyntaxKind.EqualsToken),
+                        ts.factory.createIdentifier('t')
+                    )
+                )
+            )
         ];
 
         if (details.propertyNames.length > 0) {