+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # vNext
 - Fixes issues where external interfaces were imported using absolute file paths 
   which produced builds that could not be moved
+- Fixes issues where classes/interfaces defined within functions would not be properly accessible via reflection 
+  due to the per-file type store changes.
 
 # v0.4.4